requirements.txt
# Core numerical and scientific computing
numpy>=1.22,<2.0
scipy>=1.8,<2.0

# Plotting and visualization
matplotlib>=3.5,<4.0

# Image processing for TIFF compression
Pillow>=9.0,<11.0

# Optional: For enhanced data analysis (used in some scripts)
pandas>=1.4,<3.0
